refactor(extensions): move Lark reviewer notification provider
