+++ loopx/capabilities/issue_fix/provider_hooks.py
@@ -7,16 +7,18 @@ from typing import Any
 
 InboxOperation = Callable[..., dict[str, Any]]
 InboxContainsText = Callable[..., bool]
+NotificationSinkAdapter = Callable[..., dict[str, Any]]
 
 
 @dataclass(frozen=True)
-class ReviewerInboxHooks:
+class IssueFixReviewerProviderHooks:
     """Provider-neutral operations used by issue-fix reviewer workflows."""
 
     inspect: InboxOperation
     acknowledge: InboxOperation
     contains_text: InboxContainsText
+    notification_adapter: NotificationSinkAdapter
     activation: Mapping[str, object]
 
 
-ReviewerInboxHooksFactory = Callable[[], ReviewerInboxHooks]
+IssueFixReviewerProviderHooksFactory = Callable[[], IssueFixReviewerProviderHooks]
