+++ loopx/capabilities/issue_fix/reviewer_notification.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import hashlib
-import html
 import json
 import re
 import subprocess
@@ -33,19 +32,7 @@ ISSUE_FIX_REVIEWER_NOTIFICATION_QUEUE_RECEIPT_SCHEMA_VERSION = (
 ISSUE_FIX_REVIEWER_NOTIFICATION_LEGACY_QUEUE_RECEIPT_SCHEMA_VERSION = (
     "issue_fix_reviewer_notification_queue_receipt_v0"
 )
-LARK_PERMISSION_PATTERN = re.compile(
-    r"(?:missing\s+scope|permission|not\s+in\s+(?:the\s+)?chat|"
-    r"lacks?\s+authority|99991672|230027|232033)",
-    re.IGNORECASE,
-)
-LARK_SEARCH_PERMISSION_PATTERN = re.compile(
-    r"search:message|missing\s+scope[^\n]*(?:message|search)",
-    re.IGNORECASE,
-)
 SAFE_LOCAL_KEY_PATTERN = re.compile(r"[A-Za-z0-9._-]{1,100}")
-LARK_DESTINATION_PATTERN = re.compile(r"oc_[A-Za-z0-9_-]+")
-LARK_MEMBER_PATTERN = re.compile(r"ou_[A-Za-z0-9_-]+")
-LARK_MESSAGE_PATTERN = re.compile(r"om_[A-Za-z0-9_-]+")
 LOCAL_TIME_PATTERN = re.compile(r"(?:[01]\d|2[0-3]):[0-5]\d")
 
 CommandRunner = Callable[[Sequence[str]], Mapping[str, Any]]
@@ -103,7 +90,7 @@ def load_goal_reviewer_notification_sinks_input(
     goal: Mapping[str, Any],
     project: str | Path,
 ) -> dict[str, Any] | None:
-    """Load the goal-default sink config and require explicit profile bindings."""
+    """Load a registered provider sink config without exposing its local path."""
 
     path = goal_reviewer_notification_config_path(goal=goal, project=project)
     if path is None:
@@ -124,16 +111,6 @@ def load_goal_reviewer_notification_sinks_input(
     for sink in sinks:
         if not isinstance(sink, Mapping):
             raise ValueError("goal reviewer notification sinks must be objects")
-        if sink.get("sink_kind") == "lark_chat" and not (
-            SAFE_LOCAL_KEY_PATTERN.fullmatch(str(sink.get("reader_profile") or ""))
-            and sink.get("reader_identity") == "user"
-            and SAFE_LOCAL_KEY_PATTERN.fullmatch(str(sink.get("sender_profile") or ""))
-            and sink.get("sender_identity") == "bot"
-        ):
-            raise ValueError(
-                "goal lark reviewer notification requires explicit reader/user "
-                "and sender/bot profile bindings"
-            )
     return payload
 
 
@@ -408,7 +385,7 @@ def _queued_result(
     window: Mapping[str, Any],
 ) -> dict[str, Any]:
     if not SAFE_LOCAL_KEY_PATTERN.fullmatch(sink_instance_key):
-        return _public_result(
+        return build_reviewer_notification_sink_result(
             sink_kind=sink_kind,
             reviewer_handles=[],
             idempotency_key=None,
@@ -424,7 +401,7 @@ def _queued_result(
         sink_instance_key=sink_instance_key,
         reviewer_handles=reviewer_handles,
     )
-    result = _public_result(
+    result = build_reviewer_notification_sink_result(
         sink_kind=sink_kind,
         reviewer_handles=reviewer_handles,
         idempotency_key=key,
@@ -455,62 +432,7 @@ def _queued_result(
     return result
 
 
-def _find_message_id(value: Any) -> str | None:
-    if isinstance(value, Mapping):
-        candidate = value.get("message_id")
-        if isinstance(candidate, str) and LARK_MESSAGE_PATTERN.fullmatch(candidate):
-            return candidate
-        for nested in value.values():
-            found = _find_message_id(nested)
-            if found:
-                return found
-    elif isinstance(value, list):
-        for nested in value:
-            found = _find_message_id(nested)
-            if found:
-                return found
-    return None
-
-
-def _parse_json_object(value: Any) -> Mapping[str, Any] | None:
-    try:
-        payload = json.loads(str(value or ""))
-    except json.JSONDecodeError:
-        return None
-    return payload if isinstance(payload, Mapping) else None
-
-
-def _payload_contains_text(value: Any, text: str) -> bool:
-    if isinstance(value, Mapping):
-        return any(_payload_contains_text(child, text) for child in value.values())
-    if isinstance(value, list):
-        return any(_payload_contains_text(child, text) for child in value)
-    return isinstance(value, str) and text in value
-
-
-def _lark_member_ids(value: Any) -> set[str]:
-    """Collect exact member ids from a provider response without retaining it."""
-
-    member_ids: set[str] = set()
-
-    def visit(node: Any) -> None:
-        if isinstance(node, Mapping):
-            for key, child in node.items():
-                if key in {"member_id", "open_id"} and isinstance(child, str):
-                    member_id = child.strip()
-                    if LARK_MEMBER_PATTERN.fullmatch(member_id):
-                        member_ids.add(member_id)
-                else:
-                    visit(child)
-        elif isinstance(node, list):
-            for child in node:
-                visit(child)
-
-    visit(value)
-    return member_ids
-
-
-def _public_result(
+def build_reviewer_notification_sink_result(
     *,
     sink_kind: str,
     reviewer_handles: Sequence[str],
@@ -526,6 +448,8 @@ def _public_result(
     semantic_dedupe_status: str | None = None,
     blocker: str | None = None,
 ) -> dict[str, Any]:
+    """Build the public-safe result contract returned by provider adapters."""
+
     result: dict[str, Any] = {
         "ok": ok,
         "schema_version": ISSUE_FIX_REVIEWER_NOTIFICATION_SINK_RESULT_SCHEMA_VERSION,
@@ -551,521 +475,6 @@ def _public_result(
     if semantic_dedupe_status:
         result["semantic_dedupe_status"] = semantic_dedupe_status
     return result
-
-
-def _lark_result(
-    *,
-    repo: str,
-    pr_number: int,
-    pr_url: str,
-    pr_title: str,
-    linked_issue_refs: Sequence[str],
-    reviewer_handles: Sequence[str],
-    sink: Mapping[str, Any],
-    receipts: set[str],
-    execute: bool,
-    runner: CommandRunner,
-) -> dict[str, Any]:
-    sink_kind = "lark_chat"
-    if sink.get("identity_scope") != "project_dedicated":
-        return _public_result(
-            sink_kind=sink_kind,
-            reviewer_handles=[],
-            idempotency_key=None,
-            status="gate_required",
-            ok=False,
-            external_write_authority_asserted=execute,
-            blocker="dedicated_bot_identity_required",
-        )
-
-    explicit_profile_bindings = any(
-        sink.get(field) is not None
-        for field in (
-            "reader_profile",
-            "reader_identity",
-            "sender_profile",
-            "sender_identity",
-        )
-    )
-    reader_profile = str(sink.get("reader_profile") or "").strip()
-    reader_identity = str(sink.get("reader_identity") or "").strip()
-    profile = str(sink.get("sender_profile") or sink.get("bot_profile") or "").strip()
-    sender_identity = str(sink.get("sender_identity") or "bot").strip()
-    expected_bot_name = public_safe_compact_text(
-        sink.get("bot_display_name"),
-        limit=100,
-    )
-    destination_id = str(sink.get("destination_id") or "").strip()
-    instance_key = str(sink.get("sink_instance_key") or "").strip()
-    if (
-        not SAFE_LOCAL_KEY_PATTERN.fullmatch(profile)
-        or profile.lower() == "default"
-        or not SAFE_LOCAL_KEY_PATTERN.fullmatch(instance_key)
-        or not expected_bot_name
-        or sender_identity != "bot"
-        or (
-            explicit_profile_bindings
-            and (
-                not SAFE_LOCAL_KEY_PATTERN.fullmatch(reader_profile)
-                or reader_identity != "user"
-                or not sink.get("sender_profile")
-            )
-        )
-    ):
-        return _public_result(
-            sink_kind=sink_kind,
-            reviewer_handles=[],
-            idempotency_key=None,
-            status="gate_required",
-            ok=False,
-            external_write_authority_asserted=execute,
-            blocker="dedicated_bot_identity_required",
-        )
-    if not LARK_DESTINATION_PATTERN.fullmatch(destination_id):
-        return _public_result(
-            sink_kind=sink_kind,
-            reviewer_handles=[],
-            idempotency_key=None,
-            status="gate_required",
-            ok=False,
-            external_write_authority_asserted=execute,
-            blocker="reviewer_notification_destination_unavailable",
-        )
-
-    raw_identities = sink.get("reviewer_identities")
-    identities = {
-        handle: value
-        for raw_handle, value in (
-            raw_identities.items() if isinstance(raw_identities, Mapping) else []
-        )
-        if (handle := _normalise_handle(raw_handle)) is not None
-    }
-    resolved: list[tuple[str, str, str]] = []
-    for handle in reviewer_handles:
-        value = identities.get(handle)
-        identity = value if isinstance(value, Mapping) else {}
-        member_id = str(identity.get("member_id") or "").strip()
-        display_name = public_safe_compact_text(
-            identity.get("display_name") or handle,
-            limit=80,
-        )
-        if not LARK_MEMBER_PATTERN.fullmatch(member_id):
-            return _public_result(
-                sink_kind=sink_kind,
-                reviewer_handles=[],
-                idempotency_key=None,
-                status="gate_required",
-                ok=False,
-                external_write_authority_asserted=execute,
-                blocker="reviewer_notification_identity_unresolved",
-            )
-        resolved.append((handle, member_id, display_name or handle))
-
-    key = _idempotency_key(
-        repo=repo,
-        pr_number=pr_number,
-        sink_kind=sink_kind,
-        sink_instance_key=instance_key,
-        reviewer_handles=reviewer_handles,
-    )
-    if key in receipts:
-        return _public_result(
-            sink_kind=sink_kind,
-            reviewer_handles=reviewer_handles,
-            idempotency_key=key,
-            status="already_notified",
-            ok=True,
-            external_write_authority_asserted=execute,
-            notification_verified=True,
-            semantic_dedupe_status="persisted_evidence_match",
-        )
-    if not execute:
-        return _public_result(
-            sink_kind=sink_kind,
-            reviewer_handles=reviewer_handles,
-            idempotency_key=key,
-            status="preview_ready",
-            ok=True,
-            external_write_authority_asserted=False,
-        )
-
-    reader_verified = False
-    semantic_dedupe_status = "not_configured"
-    if explicit_profile_bindings:
-        try:
-            reader_status = runner(
-                [
-                    "lark-cli",
-                    "--profile",
-                    reader_profile,
-                    "auth",
-                    "status",
-                    "--verify",
-                    "--json",
-                ]
-            )
-        except (OSError, subprocess.SubprocessError):
-            reader_status = {"returncode": 1}
-        reader_payload = _parse_json_object(reader_status.get("stdout"))
-        reader = (
-            (reader_payload.get("identities") or {}).get("user")
-            if isinstance(reader_payload, Mapping)
-            and isinstance(reader_payload.get("identities"), Mapping)
-            else {}
-        )
-        reader = reader if isinstance(reader, Mapping) else {}
-        reader_verified = bool(
-            reader_status.get("returncode") == 0
-            and reader.get("available") is True
-            and reader.get("verified") is True
-        )
-        if not reader_verified:
-            return _public_result(
-                sink_kind=sink_kind,
-                reviewer_handles=reviewer_handles,
-                idempotency_key=key,
-                status="gate_required",
-                ok=False,
-                external_write_authority_asserted=True,
-                blocker="reviewer_notification_reader_auth_required",
-            )
-        try:
-            members = runner(
-                [
-                    "lark-cli",
-                    "--profile",
-                    reader_profile,
-                    "im",
-                    "chat.members",
-                    "get",
-                    "--chat-id",
-                    destination_id,
-                    "--member-id-type",
-                    "open_id",
-                    "--page-all",
-                    "--as",
-                    "user",
-                    "--format",
-                    "json",
-                ]
-            )
-        except (OSError, subprocess.SubprocessError):
-            members = {"returncode": 1}
-        if members.get("returncode") != 0:
-            provider_error = " ".join(
-                str(members.get(field) or "") for field in ("stderr", "stdout")
-            )
-            return _public_result(
-                sink_kind=sink_kind,
-                reviewer_handles=reviewer_handles,
-                idempotency_key=key,
-                status="gate_required",
-                ok=False,
-                external_write_authority_asserted=True,
-                reader_identity_verified=True,
-                blocker=(
-                    "lark_bot_group_access_required"
-                    if LARK_PERMISSION_PATTERN.search(provider_error)
-                    else "reviewer_notification_provider_failed"
-                ),
-            )
-
-        try:
-            semantic_search = runner(
-                [
-                    "lark-cli",
-                    "--profile",
-                    reader_profile,
-                    "im",
-                    "+messages-search",
-                    "--chat-id",
-                    destination_id,
-                    "--query",
-                    pr_url,
-                    "--page-size",
-                    "20",
-                    "--as",
-                    "user",
-                    "--no-reactions",
-                    "--format",
-                    "json",
-                ]
-            )
-        except (OSError, subprocess.SubprocessError):
-            semantic_search = {"returncode": 1}
-        semantic_payload = _parse_json_object(semantic_search.get("stdout"))
-        if semantic_search.get("returncode") == 0:
-            if semantic_payload is None:
-                return _public_result(
-                    sink_kind=sink_kind,
-                    reviewer_handles=reviewer_handles,
-                    idempotency_key=key,
-                    status="gate_required",
-                    ok=False,
-                    external_write_authority_asserted=True,
-                    reader_identity_verified=True,
-                    semantic_dedupe_status="provider_failed",
-                    blocker="reviewer_notification_dedupe_readback_failed",
-                )
-            if _payload_contains_text(semantic_payload, pr_url):
-                return _public_result(
-                    sink_kind=sink_kind,
-                    reviewer_handles=reviewer_handles,
-                    idempotency_key=key,
-                    status="already_notified",
-                    ok=True,
-                    external_write_authority_asserted=True,
-                    verification_performed=True,
-                    notification_verified=True,
-                    reader_identity_verified=True,
-                    semantic_dedupe_status="configured_chat_match",
-                )
-            semantic_dedupe_status = "configured_chat_no_match"
-        else:
-            provider_error = " ".join(
-                str(semantic_search.get(field) or "") for field in ("stderr", "stdout")
-            )
-            if LARK_SEARCH_PERMISSION_PATTERN.search(provider_error):
-                semantic_dedupe_status = "permission_fallback"
-            else:
-                return _public_result(
-                    sink_kind=sink_kind,
-                    reviewer_handles=reviewer_handles,
-                    idempotency_key=key,
-                    status="gate_required",
-                    ok=False,
-                    external_write_authority_asserted=True,
-                    reader_identity_verified=True,
-                    semantic_dedupe_status="provider_failed",
-                    blocker="reviewer_notification_dedupe_readback_failed",
-                )
-
-    try:
-        identity_status = runner(
-            [
-                "lark-cli",
-                "--profile",
-                profile,
-                "auth",
-                "status",
-                "--verify",
-                "--json",
-            ]
-        )
-    except (OSError, subprocess.SubprocessError):
-        identity_status = {"returncode": 1}
-    identity_payload = _parse_json_object(identity_status.get("stdout"))
-    bot_identity = (
-        (identity_payload.get("identities") or {}).get("bot")
-        if isinstance(identity_payload, Mapping)
-        and isinstance(identity_payload.get("identities"), Mapping)
-        else {}
-    )
-    bot_identity = bot_identity if isinstance(bot_identity, Mapping) else {}
-    if not (
-        identity_status.get("returncode") == 0
-        and bot_identity.get("available") is True
-        and bot_identity.get("verified") is True
-        and str(bot_identity.get("appName") or "") == expected_bot_name
-    ):
-        return _public_result(
-            sink_kind=sink_kind,
-            reviewer_handles=reviewer_handles,
-            idempotency_key=key,
-            status="gate_required",
-            ok=False,
-            external_write_authority_asserted=True,
-            blocker="dedicated_bot_identity_mismatch",
-            reader_identity_verified=reader_verified,
-        )
-
-    if explicit_profile_bindings:
-        try:
-            sender_members = runner(
-                [
-                    "lark-cli",
-                    "--profile",
-                    profile,
-                    "im",
-                    "chat.members",
-                    "get",
-                    "--chat-id",
-                    destination_id,
-                    "--member-id-type",
-                    "open_id",
-                    "--page-all",
-                    "--as",
-                    "bot",
-                    "--format",
-                    "json",
-                ]
-            )
-        except (OSError, subprocess.SubprocessError):
-            sender_members = {"returncode": 1}
-        if sender_members.get("returncode") != 0:
-            provider_error = " ".join(
-                str(sender_members.get(field) or "") for field in ("stderr", "stdout")
-            )
-            return _public_result(
-                sink_kind=sink_kind,
-                reviewer_handles=reviewer_handles,
-                idempotency_key=key,
-                status="gate_required",
-                ok=False,
-                external_write_authority_asserted=True,
-                bot_identity_verified=True,
-                reader_identity_verified=reader_verified,
-                blocker=(
-                    "lark_bot_group_access_required"
-                    if LARK_PERMISSION_PATTERN.search(provider_error)
-                    else "reviewer_notification_provider_failed"
-                ),
-            )
-        sender_member_ids = _lark_member_ids(
-            _parse_json_object(sender_members.get("stdout"))
-        )
-        if any(member_id not in sender_member_ids for _, member_id, _ in resolved):
-            return _public_result(
-                sink_kind=sink_kind,
-                reviewer_handles=reviewer_handles,
-                idempotency_key=key,
-                status="gate_required",
-                ok=False,
-                external_write_authority_asserted=True,
-                bot_identity_verified=True,
-                reader_identity_verified=reader_verified,
-                blocker="reviewer_notification_identity_unresolved",
-            )
-
-    provider_idempotency_key = f"loopx-{key.partition(':')[2][:32]}"
-    mentions = " ".join(
-        f'<at user_id="{member_id}">{html.escape(display_name)}</at>'
-        for _, member_id, display_name in resolved
-    )
-    issue_clause = (
-        f"（修复 {', '.join(linked_issue_refs)}）" if linked_issue_refs else ""
-    )
-    summary = f"：{pr_title}" if pr_title else ""
-    content = json.dumps(
-        {
-            "text": f"{mentions} 请帮忙 review PR #{pr_number}{issue_clause}{summary}。{pr_url}"
-        },
-        ensure_ascii=False,
-        separators=(",", ":"),
-    )
-    send_args = [
-        "lark-cli",
-        "--profile",
-        profile,
-        "im",
-        "+messages-send",
-        "--chat-id",
-        destination_id,
-        "--content",
-        content,
-        "--msg-type",
-        "text",
-        "--idempotency-key",
-        provider_idempotency_key,
-        "--as",
-        "bot",
-        "--format",
-        "json",
-    ]
-    try:
-        send = runner(send_args)
-    except (OSError, subprocess.SubprocessError):
-        return _public_result(
-            sink_kind=sink_kind,
-            reviewer_handles=reviewer_handles,
-            idempotency_key=key,
-            status="gate_required",
-            ok=False,
-            external_write_authority_asserted=True,
-            blocker="reviewer_notification_provider_unavailable",
-            bot_identity_verified=True,
-            reader_identity_verified=reader_verified,
-        )
-    if send.get("returncode") != 0:
-        provider_error = " ".join(
-            str(send.get(field) or "") for field in ("stderr", "stdout")
-        )
-        return _public_result(
-            sink_kind=sink_kind,
-            reviewer_handles=reviewer_handles,
-            idempotency_key=key,
-            status="gate_required",
-            ok=False,
-            external_write_authority_asserted=True,
-            blocker=(
-                "lark_bot_group_access_required"
-                if LARK_PERMISSION_PATTERN.search(provider_error)
-                else "reviewer_notification_provider_failed"
-            ),
-            bot_identity_verified=True,
-            reader_identity_verified=reader_verified,
-        )
-
-    send_payload = _parse_json_object(send.get("stdout"))
-    message_id = _find_message_id(send_payload)
-    if not message_id:
-        return _public_result(
-            sink_kind=sink_kind,
-            reviewer_handles=reviewer_handles,
-            idempotency_key=key,
-            status="sent_unverified",
-            ok=False,
-            external_write_authority_asserted=True,
-            external_write_performed=True,
-            blocker="lark_notification_not_verified",
-            bot_identity_verified=True,
-            reader_identity_verified=reader_verified,
-        )
-    try:
-        readback = runner(
-            [
-                "lark-cli",
-                "--profile",
-                profile,
-                "im",
-                "+messages-mget",
-                "--message-ids",
-                message_id,
-                "--as",
-                "bot",
-                "--no-reactions",
-                "--format",
-                "json",
-            ]
-        )
-    except (OSError, subprocess.SubprocessError):
-        readback = {"returncode": 1}
-    readback_payload = _parse_json_object(readback.get("stdout"))
-    readback_text = (
-        json.dumps(readback_payload, ensure_ascii=False, sort_keys=True)
-        if readback_payload is not None
-        else ""
-    )
-    verified = bool(
-        readback.get("returncode") == 0
-        and message_id in readback_text
-        and pr_url in readback_text
-    )
-    return _public_result(
-        sink_kind=sink_kind,
-        reviewer_handles=reviewer_handles,
-        idempotency_key=key,
-        status="sent_verified" if verified else "sent_unverified",
-        ok=verified,
-        external_write_authority_asserted=True,
-        external_write_performed=True,
-        verification_performed=True,
-        notification_verified=verified,
-        bot_identity_verified=True,
-        reader_identity_verified=reader_verified,
-        semantic_dedupe_status=semantic_dedupe_status,
-        blocker=None if verified else "lark_notification_not_verified",
-    )
 
 
 def validate_issue_fix_reviewer_notification_sinks_result(
@@ -1321,9 +730,7 @@ def build_issue_fix_reviewer_notification_sinks_result(
         else "default_unrestricted"
     )
 
-    adapters: dict[str, NotificationSinkAdapter] = {"lark_chat": _lark_result}
-    if sink_adapters:
-        adapters.update(sink_adapters)
+    adapters = dict(sink_adapters or {})
     results: list[dict[str, Any]] = []
     for sink in sinks:
         sink_kind = str(sink.get("sink_kind") or "").strip()
@@ -1339,14 +746,14 @@ def build_issue_fix_reviewer_notification_sinks_result(
             if sink_kind and SAFE_LOCAL_KEY_PATTERN.fullmatch(sink_instance_key)
             else None
         )
-        if sink_kind == "lark_chat" and pr_url in semantic_history_pr_refs:
+        if pr_url in semantic_history_pr_refs:
             if current_key:
                 receipts.add(current_key)
         adapter = adapters.get(sink_kind)
         if adapter and execute and not delivery_window["allowed"]:
             queued_key = current_key
             if queued_key and queued_key in receipts:
-                result = _public_result(
+                result = build_reviewer_notification_sink_result(
                     sink_kind=sink_kind,
                     reviewer_handles=reviewers,
                     idempotency_key=queued_key,
@@ -1356,7 +763,7 @@ def build_issue_fix_reviewer_notification_sinks_result(
                     notification_verified=True,
                 )
             elif queued_key and queued_key in queued_receipts_by_key:
-                result = _public_result(
+                result = build_reviewer_notification_sink_result(
                     sink_kind=sink_kind,
                     reviewer_handles=reviewers,
                     idempotency_key=queued_key,
@@ -1395,7 +802,7 @@ def build_issue_fix_reviewer_notification_sinks_result(
                 runner=runner,
             )
         else:
-            result = _public_result(
+            result = build_reviewer_notification_sink_result(
                 sink_kind=public_safe_compact_text(sink_kind, limit=50) or "unknown",
                 reviewer_handles=[],
                 idempotency_key=None,
