+++ loopx/extensions/lark/__init__.py
@@ -5,3 +5,4 @@ LARK_INBOX_READ_PERMISSION = "lark.inbox.read"
 LARK_INBOX_WRITE_PERMISSION = "lark.inbox.write"
 LARK_REPLY_PERMISSION = "lark.reply.send"
 LARK_COLLECTOR_PERMISSION = "lark.collector.manage"
+LARK_REVIEWER_NOTIFICATION_PERMISSION = "lark.reviewer_notification.send"
